backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
web3==6.11.0
pydantic==2.4.2
bcrypt==4.0.1
jwt==1.3.1
PyJWT==2.8.1
requests==2.31.0
cryptography==41.0.4
gunicorn==21.2.0
